Moved raw /jit kernels Renamed kernels
