+++ src/qttools/kernels/__init__.py
@@ -1,25 +1,7 @@
 # Copyright (c) 2024 ETH Zurich and the authors of the qttools package.
 
 import qttools.kernels.operator as operator
-from qttools import xp
-
-if xp.__name__ == "numpy":
-    from qttools.kernels.numba import dsbcoo as dsbcoo_kernels
-    from qttools.kernels.numba import dsbcsr as dsbcsr_kernels
-    from qttools.kernels.numba import dsbsparse as dsbsparse_kernels
-
-elif xp.__name__ == "cupy":
-    from qttools.kernels.cuda import dsbcoo as dsbcoo_kernels
-    from qttools.kernels.cuda import dsbcsr as dsbcsr_kernels
-    from qttools.kernels.cuda import dsbsparse as dsbsparse_kernels
-
-else:
-    raise ValueError(f"Unrecognized ARRAY_MODULE '{xp.__name__}'")
-
 
 __all__ = [
-    "dsbsparse_kernels",
-    "dsbcoo_kernels",
-    "dsbcsr_kernels",
     "operator",
 ]
